Initial commit for test-run.py - test-runner.
Implement
https://blueprints.launchpad.net/tarantool/+spec/tarantool-test-runner.

This commit adds an implementation of test running framework
(test-run.py), and a simplistic interactive client for
the administartive console of tarantool (admin.py).

A prototype of the first test suite is added in
directory test/box.

"run" is a convenience symlink to test-run.py.

This commit also adds "make test" goal to the
top-level makefile.