Test runner: add tests for command line arguments.
Add a new test suite, cmd. Add a new driver for this
suite, cmdline.py. This driver reads arguments from the test,
and executes the server with these arguments. Server output
is passed on as test result.

@todo: It's not necessary to start/stop the server
for 'cmd' suite. Will be fixed when how to start/stop
server is defined in suite.ini (we'll then be
able to define a no-op start/stop routines).
@todo tarantool_silverbox output is hard to auto-verify
since it contains the current time and process id.
Either implement a way to mask out pieces of output
from the result file (preferrable), or change
tarantool to not output volatile data.

test-run.py modified: we now can preprocess suite.ini
"client" parameters, and substitute $constants in them,
e.g. $server.

@todo: provide ability to substitute $constants in .test
files as well.