A fix and a test case for Bug#726778
A fix and a test case for Bug#726778 "Gopt broke wal_dir and
snap_dir: they are no longer relative to work_dir".

Ensure that snap_dir and xlog_dir can be relative to
work_dir.
Ensure that chdir happens before we handle --init-storage.
Do not mix declarations and definitions for recovery_state.
Add a test case to args.test.