Test-runner: extend the grammar to cover Bug#729758
Allow disjunctions in SELECT WHERE clause and LIMIT
clause in SELECT, to be able to cover the problem
reported in Bug#729758.

Fix a bug with prepare_gdb arguments (remove an extra
space in -e option).

Add a simple test for Bug#729758.

The following bugs were reported in the process:

Bug#729789 "Zero limit is treated the same as no limit"
Bug#730593 "Bad data if incomplete tuple"
Bug#730613 "SELECT with a disjunction on multiple keys returns
matching tuples twice"