A fix and a test case for Bug#730593
A fix and a test case for
https://bugs.launchpad.net/tarantool/+bug/730593
"Bad data if incomplete tuple"

A partial tuple could get accepted by INSERT
command, and later on show up in all search results
on the index for which this partial tuple was missing
keys.

When inserting into a TREE index Tarantool/Box did not
validate that the tuple which is being inserted actually
has fields used in the index, and would insert instead
an "ASTERISK" (wildcard) value into the index.

Additionally, when such validation was added, it turned
out that box_raise leaves a corrupt tuple in place
since txn_abort() doesn't remove the unless it's marked
as "GHOST", and tuples were marked as "GHOST" after
index insert, not before.