Test-runner: a number of changes to improve debugging
When stopping the server under gdb, send SIGTERM
to the server, not to the terminal in which
gdb is started. This gives you the option
to actually examine gdb output before the test
has ended.

When the server crashes, and we're reading
from it in a client connection, detect EOF
and abort reading (the old code was buggy
and would busyloop indefinitely in case of
a crash).

Remove an unnecessary fork() in
start().