Lua: improve the way Lua procedures return results.
Instead of using "print" and fiber->iov
to return results to the client, change the Lua procedures
implementation to simply send to the client all return values
of a procedure. A procedure, in turn, can return tuples, integers
(and, in future, strings, Lua tables or whatever).

This allows nicely to call one procedure from another, and use its
results in Lua, and call the very same procedure from the binary
protocol, and still be able to process its results on the client.