Workaround for a valgrind warning  on 64-bit systems.
Workaround a valgrind warning in strtod on 64 bit systems.

On 64-bit systems, strtod uses an optimized version of
strcasecmp to find out whether or not a string represents
a number.

This version uses int (8 byte) comparison. If the string
in turn is less than 8 bytes, or is not aligned around 8 byte
boundary, uninitialized memory can be accessed.

Work this around with a simple test (this is lame and
incomplete, but covers our cases. In future either
this in valgrind will be gone, or our code will change,
or we add a suppression).