A fix for https://bugs.launchpad.net/tarantool/+bug/992171
A fix for https://bugs.launchpad.net/tarantool/+bug/992171
"SAVE SNAPSHOT process (dumper) does not respond to signals"

Reset the signal mask/handler after a fork to their defaults.
Make sure that when the snapshot dumper process is terminated
with a signal, we report it back to the user.

No test case since it would be hard to do a concurrent
test in our framework yet.