Get rid of txn_cleanup() & delayed cleanup.
We needed to have txn_cleanup() only because a tuple
which was read in txn was used in fiber->io_vec.
Decouple txn and network I/O and make sure
we can commit/rollback txn whenever we like, regardless
of what is given to fiber->iovec.

@todo: perhaps we need to optimize performance
of port cleanup for big selects.