Implement error checks in WAL I/O.
Do not use buffered file I/O which doesn't
give a precise account of failure, when it happens.

Instead, batch writes to WAL or SNAP files and
write batches using writev(). If writev() fails,
we can accurately evaluate which requests (those that
belong to the batch) were not written to disk.

In case of a partial write, rollback the failed requests,
and seek back to the end of the last fully written
request.

To support writev() and IO with file descriptors,
implement nio library, which provides a set of
EINTR- and error log- aware wrappers around syscalls.