A fix and a test case for Bug#911641
A fix and a test case for Bug#911641
"box.fiber.sleep() works incorrectly if a fiber is attached".

Use symmetric scheduling for attached fibers. This allows box.insert(),
box.update(), box.delete() box.fiber.sleep() in an attached
fiber just the same way as in a detached one.

Add tests.

Now it's possible to resume() a fiber created in another fiber.