Implement blueprint update-field-relative-id.
https://blueprints.launchpad.net/tarantool/+spec/update-field-relative-id
https://github.com/mailru/tarantool/wiki/UPDATE-request-implementation-%28ropes%29

*Incompatible change*

semantics of field indexes in UPDATE command conatining
more than one SET, DELETE or INSERT operation is
changed by this patch.

While before the patch, all field indexes were evaluated
before the update itself, after this patch, every
operation that changes the number of fields in a tuple
affects evaluation of field indexes of the subsequent
operation: i.e. indexes are evaluated against the "current"
tuple, as if all preceeding operations have been already
executed.

This patch changes the update implementation
to use a rope for field indexes. For details,
please refer to the comment near Update command,
and feature specification in Launchpad.

This is a prerequisite patch for UPDATE PUSH/POP
operations.