A fix and a test case for Bug#1043858
A fix and a test case for Bug#1043858
"server crash on lua stack overflow on CentOS 5.4"

Workaround a bug in libobjc2 which would mistakenly
try to catch a foreign exception in @catch (id e) frame,
when there is a @finally down the stack.

A test case is added to the test suite.

A refined test case is sent to the libobjc2 maintainer.