Fix a bug when rollback would confirm LSNs out of order.
When there was a partial write, rollback would
first confirm LSNs of requests which are only trying
to enter the queue, and only then LSNs of committed
or rolled back events.

Always put requests into the queue, even if we know
that they will be rolled back. This guarantees
a correct LSN order in confirm_lsn().