Extract cooperative IO from fiber.m.
Make all subsystems (administrative console,
replication, memcached, binary protocol) use exception-based
socket API (sio.m).

Remove all fiber members related to socket I/O.
Remove fiber->peer and fiber->cookie.
The idea of the cookie is to preserve in the WAL
the originator of the request. This information is, however,
lost when saving a snapshot. To my knowledge there is no use
of this functionality, apart from pretty-printing
the WAL.

Peer is like cookie, but is textual and is used
for logging. However, now, as long as network errors
occur outside fiber context, most of the usefulness
is lost. The only remaining case is when someone is
trying to execute a write request on a read-only port.

For all other cases, sio.m already supplies SocketError
with peer name, and this is where it's most useful.