Fiber ipc: a few more review fixes, fix a bug at destruction.
Use a large timeout value to indicate an infinite timeout,
not 0. Chris J. Date teaches us that using 0 as a special
value of infinity is a bad idea.

An incomplete fix for a bug when a channel is destroyed
in atexit handler, but has fibers blocked on it.

Minor cleanups.