Make w3c isopub.ent entity files available locally.
Make w3c isopub.ent entity files available locally,
since it is part of nobody-knows which package on Ubuntu,
and may be missing completely on other distributions.

On Ubuntu 13 even though it is part of mathml, it's not listed
properly in /etc/xml/catalog apparently.

This fixes xsltproc warnings and speeds up manual generation
significantly.