Since msgpuck is an external project, remove #include <lib/library/header.h>
msgpuck is an external project and hosts its headers within
the same directory which contains its C files.

Therefore, for msgpuck, we have to add src/lib to the header search path,
whether we want it or not.

Since it's there anyway, kill include/lib and move all
headers of all other stand-alone libraries to their respective
directories in lib/.

include_directories(src/) is a total blasphemy, and is thus
removed.