Remove Lua-MsgPack serializer hints 'array' and 'map'.
In Lua, by definition, an array is something that
is indexed from 1 and doesn't have gaps in indexes.

This can be easily checked without serializer hints.

Tuple to YAML conversion should not go through a temporary
conversion to a Lua table in the future, so it should
also not be an argument for slowing things down.