Fix a YAML printing regression introduced in the previous patch.
field->compact in Lua field used to be memset() to false.
I removed the memset since it's wasted cycles in 99% of cases,
but did not add an explicit initialization of field->compact
to LUA_TTABLE case, which uses it. By default, field->compact
must be set to false, and only for tuples it should be true.