Fix gh-69: add a dedicated severity for system errors in the log.
Rationale:
Any error which implies a system administrator action needs to be
marked explicitly in the log, so that it's easy to find and act
upon.

This change the number of log levels, so the default log level was
bumped from 4 to 5. Installations which set the log level
explicitly need to adjust their setting.