Implement #! (aka hashbang, aka shebang) (gh-184)
Allow start of tarnatool_box in #! form.

Remove script_dir.

Make all configuration options optional (primary_port
was the last one required).
Make the configuration file itself optional.

If no watcher is started for the event loop, either based on
the config file settings, or in the #! script, do not start the event loop.

Update the test runner to take the new changes into account.

Fix a bug in symbols_load() when no symbols
could be loaded when path to executable was neither
relative nor absolute (but based on PATH).

Fix a bug with ignored return value of slab_arena_create()
in tuple.cc