Fix gh-279: Tarantool crash on Mac OS X
Add pagezero-size magic to link flags on Mac OS X,
as is required by LuaJIT.
The flag was dropped in my patch which changed the linking
scheme for the main executable.
On the same token, check for return value of luaL_newstate(),
and panic if it returns NULL, this should simplify
problem analysis in the future.