Rename patch.
Always use IPROTO_REQUEST_TYPE to refer to the first member
of IPROTO_HEADER. Weed out 'code' as a synonym for type.

Rename iproto_packet to iproto_header. It's neither packet nor
header, but it's at least more header than it is packet.

Restore 'row' as a variable name for an instance of iproto_header
in all places where it is used as a single replication or
write ahead log row.