recovery: if panic_if_wal_error == false, try to find the best matching xlog.
If we can't find the closest xlog starting before the current vclock
of recovery, return the first xlog with vclcok after the search pattern, to
continue recovery, even though missing some rows,
in case when of panic_if_wal_error is false.

A test case will follow in a separate patch.