gh-791: a fix and a test case
A fix and a test case for gh-791: call() return value can't contain
a Lua map.

Treat a map as a scalar, i.e. convert it to a tuple field on return
from call.

Use a simple heuristics to distinguish a map from an array:
if the first key of the map is integer 1, it's an array.
This approach works pretty well for JSON maps, which can't have
non-string keys anyway.