log: fix a sever performance bug
Any function of Lua log module would first evaluate the call stack,
and only then invoke the logger function -- even if the text is
then sent to /dev/null because log level is too low.

Short cut the log call in case the log level is low and the message
is muted.