cfg: new cfg api, without explicit passing of cfg options
With structured configuration, i.e. a Lua table for a list of
replication peers, it is difficult to pass a configuration option
to the hook function as a single C object. Don't pass
options in, but query them from the hook function with cfg_get*
API.

As a result, all "change" hooks have the same signature now,
a step closer to the pluggable configuration API.