wal: remove special fetch callbacks
Use the standard cbus API in WAL writer.

This makes it possible to exchange arbitrary messages with
the writer and run a wholesome event loop in it.

Implement a pipelined rollback.

This patch is a pre-requisite for being able to rotate
WALs when making a snapshot, as well as for removing writers_mutex
in the WAL.

Implement tx <-> wal bus statistics