Compare UUIDs consistently
The C code uses memcmp() to compare UUIDs for server selection
in replication bootstrap, but in lua there is no way to do it,
because UUIDs there are represented as strings.

This patch fixes the C implementation so that it will behave
as if UUIDs are compared as strings.

Fixes #1598