vinyl: move vlsn from vy_sequence to tx_manager and zap the former
It turns out that env->seq->vlsn (view sequence number, needed for
taking a consistent snapshot) is updated exclusively in the tx manager
thread and used only in vy_task work functions, which are scheduled by
the manager itself. So there isn't much point in keeping this variable
in the global environment and protecting it with a mutex - instead we
can store it directly in tx_manager and simply pass it as a member of
vy_task struct.

This patch isn't intended to introduce any functional changes.

Closes #1687