vinyl: remove range locking from vy_range_compact_commit
vy_range_lock() is used in order to forbid scheduling two dump or
compact operation on the same range, so there's no point in locking and
then unlocking new ranges on compact completion, because it is run from
the same thread as the scheduler. This is a legacy of the past, when
vy_range_compact_commit() was called from a worker thread. Zap it.