Fix numerous sporadically failing tests.
Don't access uninitialized memory in ev_io_closing().

anfds array is initialized with zeroes, when it's reallocated,
but anfdmax is array size, not array last element, so
anfds[anfdmax] is garbage.

When we're "closing" an fd which has never been added
to anfds (ev_io_start()) wasn't called for it), we
must check that it's within array size boundaries first.