vinyl: cleanup vy_range->min_key initialization
We don't really need to init vy_range->min_key each time we insert it to
a tree, as we do now. It's enough to do it once in a range's lifetime -
when the range is created, i.e. on recovery or on split.

While we're at it, let's not panic if we fail to alloc a tuple for
min_key, but fail gracefully.