vinyl: do not schedule tasks if all workers are busy
If woken up by range changes, the scheduler will keep queueing tasks no
matter if there are workers available for processing them. This is
incorrect, because the overall situation may change by the time a worker
completes processing its task. This patch puts the scheduler to sleep if
all workers are busy.