vinyl: stat dump bandwidth and total
Dump bandwidth is needed for calculating the threshold on exceeding
which we should start dumping ranges (aka memory watermark). The higher
the bandwidth, the later we can start dumping w/o suffering from
transaction throttling. So we want to be very conservative about
estimating the bandwidth, meaning not to overestimate it. To achieve
that, we sample all observed bandwidth values in a histogram, and take
the worst result among all observations excluding 10% worst for a
bandwidth estimate (10% percentile). Until we get any observations,
assume the dump bandwidth to be equal to 10 MB/s, which should be fine
for initial guess.

The total number of bytes dumped since vinyl was started will be used
later to calculate write amplification and throttle timeout.