vinyl: eliminate reads on recovery
Currently, we use a rather dumb way to determine if a statement from
xlog should be committed or not on recovery - we lookup the statement in
all on-disk runs and, if found, compare its lsn. Obviously, this slows
down the recovery procedure significantly. Actually, we can avoid
reading on-disk runs altogether. All we need to do is compare the xlog
statement's lsn with the range's max lsn - we should only commit the
statement if the former is greater.

Closes #1697