vinyl: read iterator: do not update curr_stmt when jumping to next range
vy_read_iterator->curr_stmt is supposed to store the last statement
returned by vy_read_iterator_next() to be used for iterator restore.
However, currently it is also updated in vy_read_iterator_next_range().
If the iterator is restored after such an update it will skip the first
statement in the range, which might result in invalid select result. For
instance, if the first statement (the newest one) in a range is a delete
followed by an upsert for the same key (which is older), we will skip
the delete on restore, and return the deleted key, see

  vy_read_iterator_restore:
    vy_merge_iterator_restore:
      vy_run_iterator_restore:

        if (vy_stmt_compare(fnd, last_stmt->data, itr->index->key_def) == 0) {
                position_changed = false;
                if (next_stmt->lsn >= last_stmt->lsn) {
                        /* skip the same stmt to next stmt or older version */

This results in occasional vinyl/options test failures.

Closes #1853