vinyl: rework link between compaction source and dest range
On compaction we link range->mem of the destination range(s) to the
source range, so that the same ->mem can be accessed from two ranges.
This looks confusing, and this is what prevents us from using lists for
linking mems. Let's instead add dest ranges to range->compact_list of
the source range.