vinyl: always create runs from the tx thread
Currently, we create runs in vy_range_write_run(), which is called from
a worker thread on dump/compaction. In order to move vinyl metadata to
memtx, we'll need to allocate an id upon run creation, which involves
modification of the memtx table, which can only be done from the tx
thread. So let's create the run for dump/compaction in advance: when
preparing for dump/compaction.