vinyl: do not check index/range version when optimizing upserts
We abort upsert optimization if a range is dumped or compacted, even if
it is not the range the key being squashed belongs to. This is needless,
because the worst thing that can happen if we proceed to inserting the
resulting statement is that there will be one more REPLACE in the index
for the same key. That said, let's remove index/range version checking
from there.