vinyl: add in_shutdown argument to vy_task_ops->{abort,complete}
When vinyl metadata is moved to memtx, ->abort and ->complete callbacks
will need to insert/delete records from a memtx table. Since these
callbacks may be called from the engine destructor, when memtx tables
cannot be modified any more, we need a way to notify them about that.
This patch adds in_shutdown argument to the callbacks, which is set on
shutdown and unset otherwise.