vinyl: simplify vy_index_open_or_create()
On remote recovery, no matter if it's initial or final join, we must
always create the index directory, because we are receiving the initial
data set and the local directory is empty. On the contrary, on local
recovery, the index directory must have been created and optionally
populated (if there were dumps), so all we need to do is open it and try
to load index files. Anyway, there's no point in explicitly checking if
the index directory exists or not.

This fix breaks replication/hot_standby test in case of vinyl engine

It worked before since vinyl used to create an index directory
automatically at recovery if it didn't exist.

Now, as the test may create a vinyl space, we must specify vinyl_dir
along with wal_dir and snap_dir in config for hot standby to work.