tuple: remove class TupleRef
TupleRef was used in memtx to take a temporary reference on a new
tuple.
Now this reference is kept in txn_stmt.
Vinyl iterators use box_tuple_new(), which, in turn, uses tuple_bless(),
so there is no need to create a temporary reference either: the case
of sophia, which would return a tuple with 0 references, is no longer
relevant.