vinyl: abort all tasks on shutdown
Currently, we try to complete tasks on shutdown, passing in_shutdown
flag to the ->complete callback. None of the available tasks can do
anything on shutdown, because all of them need to commit the change done
by the task to the metadata log, which requires coeio, which is
unavailable at exit time, so all existing ->complete callbacks just
return -1 if in_shutdown is set in order to abort operation. When
writing this code, I forgot to set diag in this case, which results in
crash on shutdown. Anyway, let's just drop in_shtudown from ->complete
callback and always call ->abort for all pending tasks on shutdown as
per above.