vinyl: multi-level compaction
To follow the LSM tree design, group runs of the same range into levels.
The division is ephemeral, it is recomputed after each dump and not
stored in vy_range struct. The size of the first level is set to the
largest run dumped since the last compaction. The size of each
subsequent level is level_size_ratio times larger than previous one.
When the number of runs at a level exceeds max_runs_per_level, all runs
of the maximal level that exceeds the limitation as well as all runs and
mems of preceding levels are compacted. max_runs_per_level and
level_size_ratio are set in index config. By default, they equal to 2
and 10, respectively.

Closes #1821