A separate endpoint for high priority messages in tx cord. Issue #1274
Add a separate endpoint to wake up tx fibers from the wal thread.
The endpoint can be used for any purposes which do not involve
a yield. A yield is not allowed because the endpoint handler is executed
outside a fiber pool context. This is an alternative, more reliable
fix for the problem addressed by 975e92e3fdfb1c3aa0af0022247d532f5e2bb106
and filed in gh-1892, when tx thread could deadlock if there
are too many connections or messages coming from iproto thread.