vinyl: improve cache hit rate of tuple cache
 Implement tuple cache boundaries. Now if the value is the first or
the last in EQ query, it is placed into the cache with a special mark.
The next query finds the mark and understands that there is no
reason to search the value in mems/runs.

 Also add only-one-value optimization into merge iterator: it now
stops in the second next_key call if the search key is full
(and the index in unique)