Travis CI: add OS=rumprun DIST=hw target
Rumprun is a toolchain to build unikernels based on NetBSD/Rump kernel.
Compiled images can be booted on KVM/XEN and even on "bare metal".

Unfortunally, Tarantool unikernels are still not fully functional due to
bugs in libeio and LuaJIT. However, now we have a simple way to check
cross-compilation for BSD/Rumpkernel and to bake unikernels from the code:

    OS=rumprun DIST=hw packpack
    qemu-system-x86_64 -enable-kvm -kernel build/tarantool-1.7.3.265.img