vinyl: move vy_log definition from header to source file
It's too painful to have the vy_log struct in the header while keeping
the dependency list short. We already have to allocate latch instead of
embedding it in order not to include "latch.h". Since we don't need to
know the size of the structure anywhere outside vy_log.c, it should be
OK to move the definition to the source file.