vinyl: review fixes for minor cleanup in vinyl run reader
Remove the check for actual page count exceeding the one written
in page meta: the check was incorrect.
It took into account a possibl extra row in xlog transaction,
but didn't check if the log has other transaction.

The check for extra transactions is not trivial.

The check for exact match itself is not necessary for correctness,
so simply remove until xlog API lends itself more easily to such
check.

See also gh-2157