xctl: separate gc from metadata log rotation
This patch is a preparation for centralized garbage collection. It
extracts the code doing garbage collection from xctl_rotate() and
places it in a separate function xctl_collect_garbage(). The latter
takes a signature which determines the minimal age an object has to
have to be deleted - the function only removes files left from objects
that were deleted before the log received the given signature. It is
needed for making vinyl respect box.cfg.snapshot_count.