box: fix LSN assigned on final join
The number of rows sent during initial join may be greater than the LSN
of the checkpoint sent by the master, because there are rows that do not
contribute to LSN (system spaces, etc). If this happens, LSNs assigned
on final join will be greater than LSNs assigned after bootstrapping is
complete, which breaks Vinyl logic. Fix that by resetting recovery
vclock to the checkpoint LSN before getting to final join.