wal: implement journal interface in wal_writer.
During recovery from a snapshot in memtx, patch
txn_commit() to not touch the global recovery and return a dummy
LSN.
During local recovery from a transaction log, use LSN of the
record, and do not increment recovery->vclock.
During normal WAL operation, provide an implementation which
issues an LSN in WAL module.

While refactoring the WAL remove the sloppy practice of adding
an external sentinel for an internal state of an object (writer.is_active
being used as an indicator of state of writer.current_wal).
This practice got replicated in vinyl, fix it there as well.