vinyl: reset max_dump_size in compact_new()
The dump of index-wide run can start when compaction of one or more
ranges is in progress (see #2209). If we reset the max_dump_size and
compact_priority in compact_complete, there can be the following case:

   compact_prio = x      compact_prio = y > x   compact_prio = 0
    |                 |                       |    => y has lost.
    +-----------------|-----------------------+
    |                 compact                 |
    +-----------------|-----------------------+
         +------------+
         |    dump    |
         +------------+

Reset the compact_priority and max_dump_size in compact_new() and
restore the saved values on abort.

Needed for #2209