Rename xctl back to vylog
Moving vylog to the core and naming it xctl in order to store metadata
coming from all engines in it turned out to be a premature move, because
we don't exactly know what format to use, neither do we have a proper
infrastructure yet. Move it back for now. We will get back to it once we
figure out how to implement replication groups.