Export gc info to Lua
This patch renames box.internal.gc() to box.interanl.gc.run() and adds
a new Lua function box.interanl.gc.info(), which returns a table
containing the garbage collection signature and the list of tracked
checkpoints. This may be useful for debugging. Also, this information
can be used by the snapshot daemon instead of scanning the snapshot
directory.

Example output:

  tarantool> box.internal.gc.info()
  ---
  - signature: 28
    checkpoints:
    - signature: 17
      refs: 1
    - signature: 28
      refs: 0
    - signature: 42
      refs: 1
  ...