vinyl: pin in-memory trees on tx prepare
Pin each in-memory tree which is going to be modified by a transaction
on Engine::prepare, unpin it on commit or rollback. Pinned in-memory
trees can't be dumped until they are unpinned - dump/compaction task
constructors wait until pin_count drops to 0 before handing the task
over to a worker thread. Since the constructors rotate the in-memory
tree before waiting and new transaction can only go to active in-memory
trees, they won't wait forever, only until all transactions that were
started before the task was scheduled are over.

This is needed to insert statements into in-memory trees on tx prepare,
not on tx commit as per now.

Note, since vy_scheduler->checkpoint_lsn is only set after
wal_checkpoint() while we select the in-memory tree to insert a new
statement into in Engine::prepare, i.e. before writing to WAL, in order
to guarantee snapshot consistency (i.e. that no statements inserted
after WAL rotation will make it to the snapshot) we have to change the
condition triggering rotation of the in-memory tree upon insertion of a
statement from:

  mem->min_lsn <= checkpoint_lsn

to

  mem->snapshot_version != snapshot_version

(snapshot_version is increased on snapshot before wal_checkpoint())