vinyl: don't delete pinned mem on rotation
There's an optimization in vy_range_rotate_mem() which makes it delete
the active in-memory tree instead of moving it to the frozen list
provided it's empty. However, since commit 8116f2096fef ("vinyl: pin
in-memory trees on tx prepare") an empty in-memory tree may be pinned by
an ongoing transaction, in which deleting it will result in use after
free in vy_tx_write().