vinyl: fix possible possible race between range split and mem write
Follow-up ccacba28f813 ("vinyl: fix invalid dump of pinned mem").

Similarly to dump/compaction tasks, the range split task freezes the old
range's active in-memory tree and then waits for all frozen in-memory
trees to be unpinned. This is OK for dump/compaction, because they don't
need to dump all in-memory trees of the range they are working with, so
they create a new active in-memory tree instead. In contrast to them,
range split does need to dump all data that are stored or are about to
be stored in the old range, so it doesn't create a new active tree,
which will result in null ptr dereference if vy_prepare() is called
while the task is waiting on vy_range_wait_pinned(). To avoid this, we
need to first replace the old range from the index tree and only then
wait until all concurrent transactions are over.