vinyl: remove range->max_dump_size
Runs dumped for the same range should be approximately of the same size,
so it shouldn't make much difference whether we take the size of the
most recent run or max over last run_counter_per_level runs (i.e.
range->max_dump_size) for the base size of the LSM tree. Even if for
some reason a newly created run happens to be very small it shouldn't
affect compaction policy adversely - in this case we'll assume the tree
has one more level after dump, which seems to be harmless.