vinyl: introduce the concept of run slices
When we finally move to single memory tree per index (currently we
maintain one per range), dump will result in creation of a single run.
To add such runs to ranges and be able to iterate over statements that
are within a particular range, we introduce a new concept, a run slice.
A run slice is a simple object that references a run and contains begin
and end keys inherited from the range it belongs to. Runs are now not
referenced directly by ranges, instead we use slices as an intermediary.
For now the concept is oversimplified: there may only be one slice per
run, but the following patches will remove this limitation.