vinyl: make check for empty range on split more thorough
There's a sanity check in vy_range_needs_split() that assures the
resulting ranges are not going to be empty: it checks the split key
against the oldest run's min key. The check is not enough for the slice
concept, because even if the split key is > min key, it still can be <
the beginning of the slice.