vinyl: delete empty runs right away
This reverts commit a366b5bb18a ("vinyl: keep track of empty runs").

The former single memory level design required knowledge of max LSN of
each run. Since this information can't be extracted from the run file in
general (the newest key might have been deleted by compaction), we added
it to the metadata log. Since we can get an empty run (i.e. a run w/o
file on disk) as a result of compaction or dump, we had to add a special
flag to the log per each run, is_empty, so that we could store a run
record while omitting loading run file. Thanks to the concept of slices,
this is not needed any more, so we can move min/max LSN back to the
index file and remove is_empty flag from the log. This patch starts from
removing is_empty flag.