vinyl: tx_serial.test: do not append very long strings in lua.
Patch f57151941ab9abc103c1d5f79d24c48238ab39cc introduced
generation of reproduce code and dump of it to the log. But
the problem is that the code is initially generated in a big
lua string using repeated concatenation in a loop. Such a use
of lua strings is too vulnerable in terms of performance.
Avoid repeated concatenation of lua string in tx_serial.test.