vinyl: turn vy_run->slice_count into reference counter
That is, delete the run automatically once it hits 0. This simplifies
slice deletion - we don't need to check against run->slice->slice_count
and delete run manually any more. A run is usually referenced only by
slices created for it. The only exception is dump/compaction task
creating a run - it keeps the reference counter elevated until it is
completed.