vinyl: get rid of vy_log_record->signature
Each record of the metadata log has signature. Normally (except for
VY_LOG_FORGET_RUN), signature is assigned to a record when it is written
to the log (by vy_log_write()) - then it is set to the signature of the
last checkpoint. VY_LOG_FORGET_RUN records inherit the signature of the
corresponding VY_LOG_DROP_RUN.

The signature is used for two purposes:

 1. Loading snapshot. Vylog does not use separate files for snapshot and
    log, instead it appends records written after a checkpoint to the
    file created by the checkpoint. The appended records will have
    signature equal to the signature of the checkpoint while older
    records will have signatures of previous checkpoints, so we can
    filter out records corresponding to the checkpoint.

 2. Garbage collection. The box.internal.gc.run() method takes LSN of
    the oldest snapshot to save as an argument. The signature allows us
    to filter runs that were deleted before a given LSN.

Actually, #1 doesn't need a signature - instead we could write a marker
of the end of the snapshot to the vylog file, and it would be easier to
comprehend. So, the signature is really needed only by #2, but then
there's no point in appending it to each record, because garbage
collection only affects runs, i.e. only VY_LOG_DROP_RUN needs it.

So this patch removes the signature and instead introduces:

 - a new record type VY_LOG_SNAPSHOT, which is written right after the
   last record of the snapshot;

 - a new key VY_LOG_KEY_GC_LSN exclusive for VY_LOG_DROP_RUN records,
   which is set to the LSN of the last checkpoint using the deleted run.

With this patch vy_recovery_new() has two arguments, 'signature' and
'only_snapshot' flag: 'signature' is the signature of the vylog file to
load (basically, the file name); 'only_snapshot' is a flag indicating
that only records corresponding to the snapshot must be loaded. Also,
run files that should be preserved by garbage collection are now
filtered by vy_collect_garbage() using the new key (not by signature in
vy_recovery_new() as before).